Project likely done, maybe a little more polish needed but good

+++ src/main/java/zfaria/swingy/artifacts/Armor.java
@@ -1,7 +1,7 @@
 package zfaria.swingy.artifacts;
 
 public class Armor extends Artifact {
-    public Armor(int stat) {
+    public Armor(float stat) {
         super(stat);
     }
